Initial commit
Creating the resonance-react library with the following components:

`ResonanceContext` - Provides context so that the other components can consume data from Resonance
`Content` - A wrapper that groups content together
`String` - A component whose children can be replaced by a single string from Resonance
`Markup` - A components whose children can be replaced by an HTML string from Resonance. The HTML string is passed through DomPurify before being rendered as HTML.
`Attributes` - A component that can be used to replace props and HTML attributes within its children. The `children` of this component must be a function that accepts the attributes to be replaced as an object as the argument.